HID: Fix IOT issue with Motorolla KZ500 Keyboard Usecase/: 1. Pair and connect KB 2. Send inputs from KB 3. Verify inputs from KB and trackpad
Expected Result:
HID inputs are validated on DUT

Actual Result:
HID inputs are not reflected on DUT. The connection is up and
the HID inputs are visible in snoop logs but there are no actual
inputs on DUT

Reproducible:
5/5

Fix:
Motorolla KZ500 Keyboard claims support
for multiouch functionality in SDP repord descriptor, but this
functionality doesn't work because of issues in usb hid kernel driver.
To handle this keyboard, digitizer record in report descriptor is
removed and the remaining descriptor is passed to kernel usb hid
driver during connection, which interprets the keyboard as generic
keyboard instead of multitouch keyboard and normal functioning of
keyboard works fine thereafter.

Test: HID functionalities work fine with  Motorolla KZ500 Keyboard

CRs-Fixed: 3644201
Change-Id: I888b4a8997296ebbf3d9d23dd1c52ade230f984e