HFP: Sending correct service availablity and signal strength
During call if BT is turned ON,+CIND response would be updated
with no network available and no signal, for this incorrect
response some Handsfree devices or carkits may go bad state.

This fix ensures to update +CIND response with network available
and signal strength if the call is active/held/in-progress.

CRs-Fixed: 3602285
Change-Id: I49df06fcdde3b9a7c87fbfd08e9952f855061420