Le-Audio: Synchronization between transports in below cases.
Case: 1
When Avrcp Play command received, making classic device as
active and Le-audio device as in-active.

When Classic device made active stop initiated on Le transport
and immediately acked as success to MM, without sending it to
remote. Due to which MM sent start on a2dp transport, which got
failed due to already C-I-Ses existed still.

Case:2
When BREDR device connected during ongoing LE audio streaming on
Le-audio device, BREDR becoming active and Le-audio device becomes
in-active.

During this transition updated to MM about A2dp device active, on
which it initiated stop on the LE-A sessions and start on A2dp
session, even before not completing C-I-Ses terminations. This
results in saying VSC fail for A2dp start and leading to no audio.

Fix:
Case1: Always ensure to ack success to MM when remote acks back
for the Suspend request.

Case2: Always ensure le-a in-active device update to BT app only
when remote acks back for the suspend request.

This would results in terminating the C-I-Ses on le-a transport
and initiating start VSC on A2dp transport.

CRs-Fixed: 3729535
Change-Id: Ifeb75adc5c15129d09303376ba621f5bda1565a1