GATT: Prevent creating ACL if disconnected
- Prevent creating ACL for GAP (remote name) and GATT (service
  discovery) if ACL has been disconnected prior to these request.

Process queued service discovery only when link is connected
Service discovery requests for BLE and BR/EDR links are queued
when there is a service discovery already in progress.
When a link gets disconnected while processing a queued service
discovery, the service discovery gets ignored and the state
machine is not moved to idle. Hence the subsequent service
discovery requests are not processed by the stack.
This change checks whether the link is connected before
processing the queued service discovery requests.

CRs-Fixed: 3778372

Change-Id: I0dd023ff385ff9a2a0a9d90fa4e1036c569a5e56