le_audio: Adding configs for Single bud connection
During connection with buds pair, if a single bud connects first, and any use-case starts, SingleDev config was being selected. After sometime when the second bud connected, DualDev config was being selected. But the 'params' for DualDev config did not match with that of SingleDev config, leading to BT crash.
In the fix, added corresponding SingleDev config for the required DualDev config, so that when params are matched at ENABLE, crash does not occur.

CRs-Fixed: 3787790
Change-Id: Id1d760b7be428ba33217a2e075534c35d37f10ee