le_audio: Adding configs for Single bud connection
During connection with buds pair, if a single bud connects first,
and any use-case starts, 'ASE cnt 1' config was being selected.
After sometime when the second bud connected,
'ASE cnt 2' config was being selected.
But the 'params' for 'ASE cnt 2' config did not match with that of
'ASE cnt 1' config, leading to BT crash.

In the fix, added corresponding 'ASE cnt 1' config for
the required 'ASE cnt 2' config, so that when params are matched
at ENABLE, crash does not occur.

CRs-Fixed: 3829504
Change-Id: I614edd2f485ee9ee475cb150bb1f4ae830334e97