HFP: Delay incoming indicators in DSDA case.
When 2 incoming calls are in progress in DSDA case,
sending back to back indicators when inband ringtone
is enabled, call indicators send to remote are mismatched.
This is because, during inband ringtone, SCO is already opened
before call is answered. When new indicator for incoming call is
sent, we delay that indicator to be sent to remote until the
current sco is closed. So, the indictors ssequence is mismatched.

Made changes to send indicators in correct sequence only after sco
is disconnected and Headset state machine moved to connected state
from audio on state.

CRs-Fixed: 3839885
Change-Id: I794fa42f040ad9414ef3631f0ddc2f7814240609