migrate lea uuids from U to V at earlier stage
Currently, the migration of LE audio UUIDs occurs in
btif_storage_load_bonded_leaudio, which is executed after
btif_storage_load_bonded_devices. In the
btif_storage_load_bonded_devices function, the profile UUIDs are loaded
and updated to the Java layer.
As a result, when BT is turned on for the first time after an OTA, the LE audio UUIDs cannot be updated to the Java layer. Consequently, the LE audio profile fails to connect.

Change-Id: I510533ed71f388e35dac6a13d88611c1f92096d2
CRs-Fixed: 3964858